Moving assignments to CaseSelect reducer (#4014)
adding assignments to caseSelect reducer
