+++ client/app/reader/CaseSelect/CaseSelectActions.js
@@ -70,3 +70,15 @@ export const fetchAppealUsingVeteranId = (veteranId) =>
         dispatch(fetchAppealUsingVeteranIdFailed());
       });
   };
+
+export const setViewedAssignment = (vacolsId) => ({
+  type: Constants.SET_VIEWED_ASSIGNMENT,
+  payload: {
+    vacolsId
+  }
+});
+
+export const onReceiveAssignments = (assignments) => ({
+  type: Constants.RECEIVE_ASSIGNMENTS,
+  payload: { assignments }
+});
