+++ client/app/reader/CaseSelect/CaseSelectReducer.js
@@ -1,7 +1,7 @@
 import * as Constants from './actionTypes';
 import { update } from '../../util/ReducerUtil';
 
-const initialState = {
+export const initialState = {
   selectedAppealVacolsId: null,
   isRequestingAppealsUsingVeteranId: false,
   selectedAppeal: {},
@@ -12,10 +12,12 @@ const initialState = {
   },
   caseSelectCriteria: {
     searchQuery: ''
-  }
+  },
+  assignments: [],
+  assignmentsLoaded: false
 };
 
-const caseSelectReducer = (state = initialState, action = {}) => {
+export const caseSelectReducer = (state = initialState, action = {}) => {
   switch (action.type) {
   case Constants.CLEAR_CASE_SELECT_SEARCH:
     return update(state, {
@@ -79,6 +81,27 @@ const caseSelectReducer = (state = initialState, action = {}) => {
         showNoAppealsInfoMessage: { $set: false }
       }
     });
+  case Constants.RECEIVE_ASSIGNMENTS:
+    return update(state,
+      {
+        assignments: {
+          $set: action.payload.assignments
+        },
+        assignmentsLoaded: {
+          $set: true
+        }
+      });
+  case Constants.SET_VIEWED_ASSIGNMENT:
+    return update(state,
+      {
+        assignments: {
+          $apply: (existingAssignments) =>
+            existingAssignments.map((assignment) => ({
+              ...assignment,
+              viewed: assignment.vacols_id === action.payload.vacolsId ? true : assignment.viewed
+            }))
+        }
+      });
   default:
     return state;
   }
