+++ client/app/reader/CaseSelect/actionTypes.js
@@ -7,3 +7,5 @@ export const CASE_SELECTION_PAGE = 'CASE_SELECTION_PAGE';
 export const RECEIVED_NO_APPEALS_USING_VETERAN_ID = 'RECEIVED_NO_APPEALS_USING_VETERAN_ID';
 export const RECEIVE_APPEALS_USING_VETERAN_ID_SUCCESS = 'RECEIVE_APPEALS_USING_VETERAN_ID_SUCCESS';
 export const RECEIVE_APPEALS_USING_VETERAN_ID_FAILURE = 'RECEIVE_APPEALS_USING_VETERAN_ID_FAILURE';
+export const RECEIVE_ASSIGNMENTS = 'RECEIVE_ASSIGNMENTS';
+export const SET_VIEWED_ASSIGNMENT = 'SET_VIEWED_ASSIGNMENT';
