+++ client/app/reader/LoadingScreen/LoadingScreenActions.js
@@ -1,5 +1,6 @@
 import * as Constants from '../constants';
 import { collectAllTags } from '../PdfViewer/PdfViewerActions';
+import { setViewedAssignment } from '../CaseSelect/CaseSelectActions';
 
 export const onInitialDataLoadingFail = (value = true) => ({
   type: Constants.REQUEST_INITIAL_DATA_FAILURE,
@@ -14,6 +15,7 @@ export const onInitialCaseLoadingFail = (value = true) => ({
 export const onReceiveDocs = (documents, vacolsId) =>
   (dispatch) => {
     dispatch(collectAllTags(documents));
+    dispatch(setViewedAssignment(vacolsId));
     dispatch({
       type: Constants.RECEIVE_DOCUMENTS,
       payload: {
@@ -26,11 +28,6 @@ export const onReceiveDocs = (documents, vacolsId) =>
 export const onReceiveAnnotations = (annotations) => ({
   type: Constants.RECEIVE_ANNOTATIONS,
   payload: { annotations }
-});
-
-export const onReceiveAssignments = (assignments) => ({
-  type: Constants.RECEIVE_ASSIGNMENTS,
-  payload: { assignments }
 });
 
 export const onReceiveManifests = (manifestVbmsFetchedAt, manifestVvaFetchedAt) => ({
