+++ client/app/reader/reducer.js
@@ -126,8 +126,6 @@ const initialPdfSidebarErrorState = {
 };
 
 export const initialState = {
-  assignments: [],
-  assignmentsLoaded: false,
   loadedAppealId: null,
   loadedAppeal: {},
   initialDataLoadingFail: false,
@@ -250,13 +248,6 @@ export const reducer = (state = initialState, action = {}) => {
         },
         loadedAppealId: {
           $set: action.payload.vacolsId
-        },
-        assignments: {
-          $apply: (existingAssignments) =>
-            existingAssignments.map((assignment) => ({
-              ...assignment,
-              viewed: assignment.vacols_id === action.payload.vacolsId ? true : assignment.viewed
-            }))
         }
       }
     ));
@@ -287,16 +278,6 @@ export const reducer = (state = initialState, action = {}) => {
         }
       }
     ));
-  case Constants.RECEIVE_ASSIGNMENTS:
-    return update(state,
-      {
-        assignments: {
-          $set: action.payload.assignments
-        },
-        assignmentsLoaded: {
-          $set: true
-        }
-      });
   case Constants.RECEIVE_APPEAL_DETAILS:
     return update(state,
       {
